splitting up files, new core.server package, minor cleanup
