+++ colossus/src/main/scala/colossus/core/Connection.scala
@@ -4,6 +4,7 @@ package core
 import scala.concurrent.duration._
 import java.nio.channels.{SelectionKey, SocketChannel}
 import java.net.InetSocketAddress
+import server._
 
 /**
  * Represent the connection state.  NotConnected, Connected or Connecting.
