+++ colossus/src/main/scala/colossus/core/WorkerItem.scala
@@ -7,30 +7,6 @@ import scala.concurrent.duration.FiniteDuration
 
 class WorkerItemException(message: String) extends Exception(message)
 
-/**
- * Represents the binding of an item to a worker
- *
- * @param id the id used to reference the worker item
- * @param worker the worker the item is bound to
- */
-class Context(val id: Long, val worker: WorkerRef) {
-
-  def !(message: Any)(implicit sender: ActorRef) {
-    worker.worker ! WorkerCommand.Message(id, message)
-  }
-
-  def unbind() {
-    worker.worker ! WorkerCommand.UnbindWorkerItem(id)
-  }
-
-  private[colossus] lazy val proxy : ActorRef = {
-    _proxyExists = true
-    worker.system.actorSystem.actorOf(Props(classOf[WorkerItemProxy], id, worker))
-  }
-
-  private var _proxyExists = false
-  private[colossus] def proxyExists = _proxyExists
-}
 
 /**
  * This trait contains event handler methods for when a worker item is
